UI/requirements.txt
Flask==3.0.0
werkzeug==3.0.1
google-generativeai==0.3.1
